refactor(facebook and google login): refactor and redesing google and facebook login
create a component for facebook login from the container. Change style of buttons for google and
facebook login in i=order to new style
