+++ package.json
@@ -49,7 +49,7 @@
 		"react-facebook": "8.0.4",
 		"react-facebook-login": "4.1.1",
 		"react-google-login": "5.0.5",
-		"react-icons": "3.7.0",
+		"react-icons": "^3.7.0",
 		"react-images-upload": "1.2.7",
 		"react-leaflet": "2.7.0",
 		"react-leaflet-search": "2.0.1",
