+++ src/components/helpers/validators/min-max-length-validators.js
@@ -1,7 +1,3 @@
-
-
-
-
 const minLength = min => value =>
   value && value.length < min ? `Must be ${min} characters or more` : undefined;
 const maxLength = max => value =>
