+++ src/components/login/facebook-login/index.js
@@ -1,0 +1,3 @@
+import LoginFacebook from "./facebook-login";
+
+export default LoginFacebook;
